:sparkles: #27 /api/v1/test/로 접근시 인증 요구
