+++ src/main/java/com/example/SEENEMA/config/WebSecurityConfig.java
@@ -1,0 +1,16 @@
+package com.example.SEENEMA.config;
+
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+
+@EnableWebSecurity
+public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
+    @Override
+    protected void configure(HttpSecurity http) throws Exception{
+        http.httpBasic()
+                .and()
+                .authorizeRequests()
+                .antMatchers("/api/v1/test/**").hasRole("USER");
+    }
+}
