+++ src/main/java/com/example/SEENEMA/jwt/JwtTokenProvider.java
@@ -1,0 +1,40 @@
+package com.example.SEENEMA.jwt;
+
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.SignatureAlgorithm;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.stereotype.Component;
+
+import javax.annotation.PostConstruct;
+import java.util.Base64;
+import java.util.Date;
+import java.util.List;
+
+@RequiredArgsConstructor
+@Component
+public class JwtTokenProvider {
+    private String secretKey = "seenema";
+    //토큰 유효시간 24시간
+    private long tokenValidTime = 24 * 60 * 60 * 1000L;
+    private final UserDetailsService userDetailsService;
+    //객체 초기화, secretKey 인코딩
+    @PostConstruct
+    protected void init(){
+        secretKey = Base64.getEncoder().encodeToString(secretKey.getBytes());
+    }
+
+    // Jwt 토큰 생성
+    public String createToken(String userPk, List<String> roles){
+        Claims claims = Jwts.claims().setSubject(userPk); // JWT payload에 저장되는 정보단위
+        claims.put("roles", roles); // key / value 쌍으로 저장
+        Date now = new Date();
+        return Jwts.builder()
+                .setClaims(claims)
+                .setIssuedAt(now)
+                .setExpiration(new Date(now.getTime() + tokenValidTime))
+                .signWith(SignatureAlgorithm.HS256, secretKey) // 암호화 알고리즘, signature에 들어가는 secretKey 값
+                .compact();
+    }
+}
