+++ src/main/java/com/example/SEENEMA/test/service/UserService.java
@@ -3,5 +3,5 @@ package com.example.SEENEMA.test.service;
 import org.springframework.stereotype.Service;
 
 @Service
-public class UserServiceImpl implements UserService{
+public interface UserService {
 }
